Add test_account_hash
The account hashing function is not exposed in public interface of any
of the Solana crates so when we need to use it in the light client we
need to copy the code.  Copying may introduce bugs so it’s crucial to
test other implementations.

Add a test_account_hash which calculates account hash with fixed
values.  The main purpose of this test is that other implementations
of account hashing can copy it over to verify their code.